Add num jugadores and edades filters through checkboxes #37
